added clear button
