+++ client/src/assets/scss/_clearButton.scss
@@ -1,0 +1,21 @@
+.clear-button {
+  font-size: 2.2em;
+  color: $black;
+  background-color: $yellow;
+  border-radius: 50%;
+  box-shadow: 0 0 0 2px $black;
+  margin-left: 60px;
+  margin-top: 70px;
+  transition: background-color 0.25s ease-out, color 0.25s ease-out,
+    box-shadow 0.25s ease-out;
+}
+
+.button.clear-button {
+  max-width: 40px;
+  padding: 40px;
+}
+
+.clear-button:hover {
+  color: $yellow;
+  box-shadow: 0 0 0 2px $yellow;
+}
