+++ client/src/assets/scss/main.scss
@@ -30,9 +30,10 @@ $pink: #da346e;
 @import "./rhythmicUnit";
 @import "./subdivisionDisplay";
 @import "./goButton";
+@import "./clearButton";
 
 body {
-  background-image: url("../images/background.jpeg");
+  background-image: url("../images/bg.jpeg");
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center 20px;
