+++ client/src/components/layout/ClearButton.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+const ClearButton = (props) => {
+  return (
+    <div className="button clear-button" onClick={props.handleClear}>
+      Clear
+    </div>
+  )
+}
+
+export default ClearButton
